Added some configuration options to the "App Board" projects.
Note : The "Release" version of the Quad Relay Output Board does not work because the include path is empty.  This needs to be fixed.

Note 2 : The "contingency scan" gives errors for the Quad Relay Output Board.  However, there are #define guards in the code that allows this project to compile properly.  This should also be fixed.