Changing the version numbers.  Version 1.6.2 is coming out soon.  Version 2.0.0 is as well.  I'm calling them both Build 25.
The peer assignment structure is about to change, so both ON_VERSION and ON_PARAM_VERSION need to change.  In addition, we need to be able to differentiate 2.x generation builds and devices from 1.x generation builds, so I'm starting a new trend.  Version 2.0 and beyond will have a most significant bit of value 1.  Lower versions will have a most significan bit of 0.  The other 7 bits should be the same.