Adding #define guard for function one_net_master_change_peer_data_rate.  If you are not implementing _PEER, you don't need it.  Also, apparently the Eval Board master never calls this function anyway, so I'm taking it out altogether for the master eval boards.
However, I am leaving the code where the client eval boards handle this message.  They'll never receive this message from a master eval board.  However, someone may decide to implement this on their own or an Eval Board may be used as a client where the master is NOT an Eval board, so there are possibilities where a client may need this code.