Changing PID from 6 raw bits to 12 raw bits.  Encoded PID is now 2 bits, not 1.  Raw PID bits are now arranged this way...
NNNNMSPPPPPP

NNNN is a value from 0 to 15 -- It is the number of encrypted blocks in the payload.  Each block is 8 bytes.
M is 1 if the packet is multi-hop, 0 otherwise.
S is 1 is the packet is stay-awake, 0 otherwise.
PPPPPP is a value from 0 to 0x3F containing the raw packet type.  Note that raw packet types no longer have the prefixes or suffixes of "extended", "large", "multi", "stay-awake".  "Multi" and "Syay-awake" have their own bits.  "Extended" and "large" deal with the number of blocks.  Since the size is now stored, they are no longer needed, so an "extended" packet is now a single packet with length 3.  Instead of having separate types, we now have the same type, along with a size.